Reorganize grad_slice and interval as well
