+++ fidget/src/jit/x86_64/mod.rs
@@ -6,4 +6,6 @@ pub const IMM_REG: u8 = 0;
 pub const OFFSET: u8 = 4;
 
 pub mod float_slice;
+pub mod grad_slice;
+pub mod interval;
 pub mod point;
